Add messages and comments to views

+++ board/views.py
@@ -27,6 +27,7 @@ def home(request):
         'total_archived': total_archived,
         'total_users': total_users,
     }
+    # returns homepage
     return render(request, "index.html", context)
 
 
@@ -36,12 +37,16 @@ def new_task(request):
     Get renders a blank task and profile (blank or prefilled)
     Post method creates a task and updates profile if needed.
     '''
+    # obtains blank TaskForm
     form = TaskForm()
+    # gets Profile if it exists, otherwise sets profile to None
     if Profile.objects.filter(person=request.user).exists():
         profile = get_object_or_404(Profile, person=request.user)
     else:
         profile = None
+    # POST request
     if request.method == "POST":
+        # obtains information from the form
         form = TaskForm(request.POST)
         # obtain final_date, change type and format to compare to today
         final_date_str = form.data['final_date']
@@ -54,7 +59,9 @@ def new_task(request):
             if final_date < today:
                 messages.warning(request, "Deadline cannot be in the past")
             else:
+                # calls task processing if date is at least today
                 return task_processing(request, profile, form)
+        # calls task processing if there is no deadline
         else:
             return task_processing(request, profile, form)
     # if statement sets ProfileForm depending if it already exists or not
@@ -62,6 +69,7 @@ def new_task(request):
         profile_form = ProfileForm(instance=profile)
     else:
         profile_form = ProfileForm()
+    # GET request, context contains forms
     context = {
         'form': form,
         'profile_form': profile_form,
@@ -90,7 +98,10 @@ def task_processing(request, profile, form):
         profile_form.instance.person = request.user
         profile_form.save()
         # returns user's own task list
-        return list_own_tasks(request)      
+        messages.success(request, "Task successfully created!")
+        return list_own_tasks(request)
+    # error message displayed if forms aren't valid.
+    messages.error(request, "Something went wrong...")
 
 
 def get_task_list(request):
@@ -98,6 +109,7 @@ def get_task_list(request):
     Function filters for Published tasks and obtains all profiles.
     Pagination displays number of pages depending on the number of tasks.
     '''
+    # filtering tasks and obtaining profiles
     task_list = Task.objects.filter(status="Published")
     profiles = Profile.objects.all()
     # paginator logic copied from:
@@ -122,18 +134,25 @@ def show_task(request, task_id):
     '''
     task = get_object_or_404(Task, id=task_id)
     owner_location = Profile.objects.get(person=task.owner).location
-    if request.method == "POST":
-        task.helper = request.user
-        task.status = "Ongoing"
-        task.save()
-        messages.success(request, "Task successfully accepted")
-        return home(request)
-    context = {
-        'id': task_id,
-        'task': task,
-        'owner_location': owner_location,
-    }
-    return render(request, "show_task.html", context)
+    # if there's no helper, user can apply.
+    if task.helper == None:
+        # POST request, user is saved as helper
+        if request.method == "POST":
+            task.helper = request.user
+            task.status = "Ongoing"
+            task.save()
+            messages.success(request, "Task successfully accepted")
+            return home(request)
+        # GET request 
+        context = {
+            'id': task_id,
+            'task': task,
+            'owner_location': owner_location,
+        }
+        return render(request, "show_task.html", context)
+    # error display if there is associated helper.
+    messages.error(request, "You don't seem to have access to this action.")
+    return redirect(list_own_tasks)
 
 
 @login_required
@@ -143,18 +162,27 @@ def edit_task(request, task_id):
     Post method saves the updated task and updates time.
     '''
     task = get_object_or_404(Task, id=task_id)
-    form = TaskForm(instance=task)
-    if request.method == "POST":
-        form = TaskForm(request.POST, instance=task)
-        instance = form.save(commit=False)
-        if form.is_valid():
-            instance.updated_date = datetime.now()
-            instance.save()
-        messages.success(request, "Task updated successfully!")
-        # redirects to show_task as edit is only available when there is no helper associated
-        return redirect(show_task, task_id)
-    context = {'form': form}
-    return render(request, "edit_task.html", context)
+    # owner can edit request is there is no helper.
+    if request.user == task.owner and task.helper == None:
+        form = TaskForm(instance=task)
+        # POST request
+        if request.method == "POST":
+            form = TaskForm(request.POST, instance=task)
+            instance = form.save(commit=False)
+            if form.is_valid():
+                instance.updated_date = datetime.now()
+                instance.save()
+                messages.success(request, "Task updated successfully!")
+                # redirects to show_task as edit is only available when there is no helper associated.
+                return redirect(show_task, task_id)
+            # error display if form isn't valid.
+            messages.error(request, "Something went wrong")
+        # GET request
+        context = {'form': form}
+        return render(request, "edit_task.html", context)
+    # error message if user isn't owner or helper exists.
+    messages.error(request, "You don't seem to have access to this action.")
+    return redirect(list_own_tasks)
 
 
 @login_required
@@ -164,14 +192,20 @@ def delete_task(request, task_id):
     Post method deletes the task.
     '''
     task = get_object_or_404(Task, id=task_id)
-    if request.method == "POST":
-        if request.user == task.owner:
+    # check if user is owner and there is no helper.
+    if request.user == task.owner and task.helper == None:
+        # POST request
+        if request.method == "POST":
             task.delete()
             messages.success(request, "Task deleted!")
             # returns user's list of task for a secondary deletion confirmation
-            return list_own_tasks(request)
-    context = {'task': task}
-    return render(request, "delete_task.html", context)
+            return redirect(list_own_tasks)
+        # GET request
+        context = {'task': task}
+        return render(request, "delete_task.html", context)
+    # error display if user isn't owner or helper exists.
+    messages.error(request, "You don't seem to have access to this action.")
+    return redirect(list_own_tasks)
 
 
 @login_required
@@ -185,6 +219,7 @@ def list_own_tasks(request):
     own_tasks = Task.objects.filter(owner=request.user).filter(
         status__in=["Published", "Ongoing"]
         )
+    # filter for tasks where user is the helper
     helper_tasks = Task.objects.filter(helper=request.user, status="Ongoing")
     profiles = Profile.objects.all()
     context = {
@@ -201,31 +236,47 @@ def show_ongoing_task(request, task_id):
     Get method displays the task, comment form and existing comments if any.
     Post method adds a comment and shows empty comment form again.
     '''
+    # empty comment form
     form = CommentForm()
     task = get_object_or_404(Task, id=task_id)
     owner_details = Profile.objects.filter(person=task.owner).values()
+    # obtaining all comments
     comments = task.comments.all()
-    if request.method == "POST":
-        form = CommentForm(request.POST)
-            # below lines are a customized code obtained here:
-            # https://www.youtube.com/watch?v=zJWhizYFKP0
-        instance = form.save(commit=False)
-        if form.is_valid():
-            # instance saves comment author and associated task
-            instance.author = request.user
-            instance.post = task
-            instance.save()
-            form = CommentForm()
-            messages.success(request, "Your comment was posted!")
-            return redirect(show_ongoing_task, task_id)
-    context = {
-        'id': task_id,
-        'task': task,
-        'form': form,
-        'comments': comments,
-        'owner_details': owner_details,
-    }
-    return render(request, "show_ongoing_task.html", context)
+    # check if helper exists
+    if task.helper != None:
+        # check if user is owner or helper
+        if request.user == task.owner or request.user == task.helper:
+            # POST request
+            if request.method == "POST":
+                form = CommentForm(request.POST)
+                    # below lines are a customized code obtained here:
+                    # https://www.youtube.com/watch?v=zJWhizYFKP0
+                instance = form.save(commit=False)
+                if form.is_valid():
+                    # instance saves comment author and associated task
+                    instance.author = request.user
+                    instance.post = task
+                    instance.save()
+                    form = CommentForm()
+                    messages.success(request, "Your comment was posted!")
+                    return redirect(show_ongoing_task, task_id)
+                # error display if form isn't valid.
+                messages.error(request, "Something went wrong...")
+            # GET request
+            context = {
+                'id': task_id,
+                'task': task,
+                'form': form,
+                'comments': comments,
+                'owner_details': owner_details,
+            }
+            return render(request, "show_ongoing_task.html", context)
+        # error display if user isn't owner or helper
+        messages.error(request, "You don't have access to this request.")
+        return redirect(list_own_tasks)
+    # error display in case helper doesn't exist.
+    messages.error(request, "There's nothing here...")
+    return redirect(list_own_tasks)
 
 
 @login_required
@@ -235,18 +286,29 @@ def archive_task(request, task_id):
     Post method updates the task as archived.
     '''
     task = get_object_or_404(Task, id=task_id)
-    context = {
-        "task": task,
-        "id": task_id
-    }
-    if request.method == "POST":
-        if request.user == task.owner:
-            # status change
-            task.status = "Archived"
-            task.save()
-            messages.success(request, "Glad it got sorted!")
-            return redirect(list_own_tasks)
-    return render(request, "archive_task.html", context)
+    # ownership check
+    if request.user == task.owner:
+        # POST request
+        if request.method == "POST":
+            # check if helper exists
+            if task.helper != None:
+                # status change
+                task.status = "Archived"
+                task.save()
+                messages.success(request, "Glad it got sorted!")
+                return redirect(list_own_tasks)
+            # error display in case helper exists
+            messages.error(request, "You don't seem to have access to this action.")
+            return redirect(show_task, task_id)
+        # GET request
+        context = {
+            "task": task,
+            "id": task_id
+        }
+        return render(request, "archive_task.html", context)
+    # error display if user isn't the owner
+    messages.error(request, "You don't seem to have access to this action.")
+    return redirect(list_own_tasks)
 
 
 @login_required
@@ -256,10 +318,9 @@ def filter_category(request):
     Post method displays tasks based on filtering.
     '''
     categories = Category.objects.all()
-    context = {
-        "categories": categories
-    }
+    # POST request
     if request.method == "POST":
+        # category and post filtering
         published_tasks = Task.objects.filter(status="Published")
         profiles = Profile.objects.all()
         filtered_category = request.POST.get("category-filter")
@@ -267,6 +328,7 @@ def filter_category(request):
             filtered_tasks = published_tasks.filter(category=filtered_category)
             # used to render back-up button if more than 3 tasks exist
             filtered_tasks_count = filtered_tasks.count()
+            # used to render filtered tasks depending on the category
             context = {
                 "categories": categories,
                 "filtered_tasks": filtered_tasks,
@@ -274,6 +336,12 @@ def filter_category(request):
                 "profiles": profiles,
             }
             return render(request, "filter_category.html", context)
+        # error display if user filters for none-category
+        messages.error(request, "Please choose a category")
+    # GET request
+    context = {
+        "categories": categories
+    }
     return render(request, "filter_category.html", context)
 
 
@@ -283,10 +351,12 @@ def edit_profile(request):
     Get method displays a blank or pre-filled form if exists.
     Post method creates/edits it.
     '''
+    # check for existing profile
     if Profile.objects.filter(person=request.user).exists():
         profile = get_object_or_404(Profile, person=request.user)
     else:
         profile = None
+    # POST request
     if request.method == "POST":
         if profile is not None:
             form = ProfileForm(request.POST, instance=profile)
@@ -300,11 +370,14 @@ def edit_profile(request):
             messages.success(request, "Profile updated")
             context={"form": form}
             return render(request, "profile.html", context)
+        # error display in case form isn't valid
         messages.error(request, "Error: Please try again")
+    # renders existing or empty profile, if one doesn't exist
     if profile is not None:
         form = ProfileForm(instance=profile)
     else:
         form = ProfileForm()
+    # GET request
     context = {
         "form": form,
     }
